+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ### [Unreleased]
-*
+* Fix for Insert-methods when using an anonymous type without id
 
 ### [1.7.0] - 2017-08-13
 * Full-text search